use size_t to match the argument of operator[]

+++ psi4/src/psi4/libmoinfo/moinfo.h
@@ -140,12 +140,12 @@ class MOInfo : public MOInfoBase {
     intvec get_vir() const { return (vir); }
     intvec get_all() const { return (all); }
 
-    int get_mopi(int i) const { return (mopi[i]); }
-    int get_focc(int i) const { return (focc[i]); }
-    int get_docc(int i) const { return (docc[i]); }
-    int get_actv(int i) const { return (actv[i]); }
-    int get_extr(int h) const { return (extr[h]); }
-    int get_fvir(int i) const { return (fvir[i]); }
+    int get_mopi(size_t i) const { return (mopi[i]); }
+    int get_focc(size_t i) const { return (focc[i]); }
+    int get_docc(size_t i) const { return (docc[i]); }
+    int get_actv(size_t i) const { return (actv[i]); }
+    int get_extr(size_t h) const { return (extr[h]); }
+    int get_fvir(size_t i) const { return (fvir[i]); }
 
     // Mapping functions
     intvec get_focc_to_mo() const { return (focc_to_mo); }
